issue #20 uint vs int for tree indices and sizes
Is there a reason the CBOR uses int (cose major type 1) in preference to
uint ? If not, I suggest we use uint to preserve the full range of
positive values possible.

note: I may need to update the dbor-diag output. If this change is
acceptable in principal I am happy to go do that as a requirement to
merge.