re issue#20 validity period
This PR seeks to clearly deliniate the validity period of the signature
over the receipt from any potential validity period concept for the
proof material itself.

And particularly, to avoid a situation where VDS specifications make it
impossible to express a signature validity period because it has
"camped" on the appropriate COSE headers for doing so.